+++ src/providers/DevicesProvider/AudioInputProvider.tsx
@@ -68,6 +68,9 @@ const AudioInputProvider: React.FC = ({ children }) => {
             'Previously selected audio input lost. Selecting a default device.'
           );
           meetingManager.selectAudioInputDevice(newAudioInputs[0].deviceId);
+
+          // Safari and Firefox don't have this "default" as device Id
+          // Only Chrome have this "default" device
         } else if (selectedInputRef.current === 'default') {
           console.log(
             `Audio devices updated and "default" device is selected. Reselecting input.`