+++ CHANGELOG.md
@@ -10,17 +10,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-- Fix the issue that Amazon Voice Focus does not get applied on new devices mid-meeting
+
+- Fix the issue that Amazon Voice Focus does not get applied on new devices mid-meeting.
+- Fix the issue where we call `meetingManager.leave` an additional time when we call `meetingManager.leave`.
+- Remove the observer in `LocalVideoProvider` when it is unmounted to fix memory leak.
 
 ### Added
-- Add `activeSpeakerPolicy` and `videoUplinkBandwidthPolicy` in `MeetingManagerConfig` to allow builders to pass in 
-  custom policies.
+
+- Add `activeSpeakerPolicy` and `videoUplinkBandwidthPolicy` in `MeetingManagerConfig` to allow builders to pass in custom policies.
 - For more flexibility, allow passing `MeetingManagerConfig` to `meetingManager.join` method. Passing the config here would override config passed through `MeetingProvider` props.
+- Add more details in the `AudioInputProvider` on storybook.
+- Add `MeetingStatus.Left` and set it when explicitly leaving the meeting.
+- Publish `MeetingStatus.Failed` when `audioVideoDidStop` gets triggered with one of the Failure types of `MeetingSessionStatus`.
+- Add `Terminal Failure` Meeting Status.
 
 ### Changed
+
 - Remove the audio video observers in the `audioVideoDidStop()` function instead of `leave()` function in the `MeetingManager`.
 
 ### Removed
+
+- Remove setting the `MeetingStatus` to `MeetingStatus.Loading` when we call `meetingManager.leave`.
 
 ## [2.9.1] - 2021-09-02